add update listener for entity
