+++ custom_components/view_assist/sensor.py
@@ -1,14 +1,18 @@
+import logging
 from typing import Any
 
 import voluptuous as vol
 
 from homeassistant.components.sensor import SensorEntity
-from homeassistant.core import HomeAssistant
+from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers import entity_platform
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.config_validation import make_entity_service_schema
+from homeassistant.helpers.dispatcher import async_dispatcher_connect
 
-from . import VAConfigEntry
+from .const import DOMAIN, VAConfigEntry
+
+_LOGGER = logging.getLogger(__name__)
 
 
 async def async_setup_entry(
@@ -64,6 +68,21 @@ class ViewAssistSensor(SensorEntity):
         )  # Optional for audio_only
         self._browser_id = config.data.get("browser_id")  # Optional for audio_only
         self._attr_native_value = ""
+
+    async def async_added_to_hass(self) -> None:
+        """Run when entity is about to be added to hass."""
+        self.async_on_remove(
+            async_dispatcher_connect(
+                self.hass,
+                f"{DOMAIN}_{self.config.entry_id}_update",
+                self.update,
+            )
+        )
+
+    @callback
+    def update(self, *args):
+        """Update entity."""
+        self.schedule_update_ha_state(True)
 
     @property
     def extra_state_attributes(self) -> dict[str, Any]:
